move DeviceManager into its own module
